safari-600.7-branch/LayoutTests/media/csp-blocks-video.html

<meta http-equiv="Content-Security-Policy" content="media-src 'none'">
<video controls></video>
<script src=media-file.js></script>
<script src=video-test.js></script>
<p>This test passes if it doesn't alert failure.</p>
<script>
var mediaFile = findMediaFile("video", "content/test");
video.src = mediaFile;

waitForEvent('loadedmetadata', function () { 
    alert('FAIL');
    endTest();
} );
video.load();
addEventListener('load', endTest, false);
</script>
