LayoutTests/platform/chromium/media/csp-blocks-video-expected.txt

CONSOLE MESSAGE: line 1: Refused to load media from 'test.ogv' because of Content-Security-Policy.

END OF TEST
This test passes if it doesn't alert failure.
